fix(job): refactor job update functions to streamline synchronization and enhance readability
